fix bloc for counter screen
